requirements.txt
Jinja2==2.11.1
python-dotenv==0.10.3